requirements.txt
arrow==1.3.0
asgiref==3.7.2
astroid==3.1.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.2.0
bleach==6.1.0
blinker==1.7.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1677499911949/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1687884021435/work
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.38.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1689351751701/work
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1669907009957/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1685101166527/work
cookiecutter==2.6.0
coverage==7.4.1
crispy-bootstrap5==2024.2
cryptography==42.0.4
defusedxml==0.7.1
dill==0.3.8
dj-database-url==0.5.0
dj-rest-auth==6.0.0
Django==4.2.10
django-allauth==0.61.1
django-browser-reload==1.12.1
django-cloudinary-storage==0.3.0
django-crispy-forms==2.1
django-environ==0.11.2
django-filter==24.2
django-summernote==0.8.20.0
django-tailwind==3.6.0
django-tailwindcss==0.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Flask==3.0.2
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
gunicorn==20.1.0
idna==3.6
importlib-metadata==7.0.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1632759296524/work
jsonpointer==2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.0.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1666836642684/work
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==22.0.0
pyparsing==3.1.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.8
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.46
sqlparse==0.4.4
tailwind==3.1.5b0
tailwindcss==0.0.1
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.4
toolz==0.11.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1677948868469/work
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
zipp==3.17.0
zstandard==0.19.0
